Split apart root __init__.py into smaller modules
* InfrahubClient -> client.py
* InfrahubNode, Attribute, RelatedNode & RelationshipManager -> node.py
* BranchData, RFileData etc -> data.py
